Some fixes to owner administration
Fixed my SQLAlchemy relationships (oops!).
Since owned is now a list, I'm just running through a wrapper function.
Might've been worth making it a one-to-one relationship, but not sure
where I'm going with the whole groups thing (can people be members of
multiple groups in the future?). Wrapper works for now.

Also fixed a bug in the template.